Clean up _allToken checks...
Since this code is checking true / false and then inserting bool literal true / false respecitvely, we can just use the variable.

the only other differnce between the 2 branches of the conditiional is to ensure _tokenId is 0 when _allTokens is true. We can handle that with a conditional assignment of 0, or keeping the existing value on _tokenId.