Refactor unneeded if statements prior to loops...
When we are looping over the something based on the length (how many hashes), we don't need to check if it's greater than 0... If it is 0, the for loop won't execute at all anyway.

This also simplified the logic and readability of the return pattern: just return from in the loop if true, and otherwise return false.

Eliminating the status variable reduces storage and complexit, and improves readability.